Extract test coverage utils into their own package
