+++ e2e-tests/ui-components/package.json
@@ -10,6 +10,7 @@
     "@evervault/ui-components": "workspace:*"
   },
   "devDependencies": {
+    "@repo/test-coverage": "workspace:*",
     "v8-to-istanbul": "^9.3.0"
   }
 }
